requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.2.0
python-multipart==0.0.6
openpyxl==3.1.2 